+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_api_impls.dart
@@ -563,7 +563,7 @@ class JavaScriptChannelFlutterApiImpl extends JavaScriptChannelFlutterApi {
         .getInstanceWithWeakReference(instanceId) as JavaScriptChannel?;
     assert(
       instance != null,
-      'InstanceManager does not contain an JavaScriptChannel with instanceId: $instanceId',
+      'InstanceManager does not contain a JavaScriptChannel with instanceId: $instanceId',
     );
     instance!.postMessage(message);
   }
@@ -617,11 +617,11 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
-      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
-      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
     );
     if (instance!.onPageFinished != null) {
       instance.onPageFinished!(webViewInstance!, url);
@@ -636,11 +636,11 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
-      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
-      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
     );
     if (instance!.onPageStarted != null) {
       instance.onPageStarted!(webViewInstance!, url);
@@ -661,11 +661,11 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
-      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
-      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
     );
     // ignore: deprecated_member_use_from_same_package
     if (instance!.onReceivedError != null) {
@@ -691,11 +691,11 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
-      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
-      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
     );
     if (instance!.onReceivedRequestError != null) {
       instance.onReceivedRequestError!(
@@ -718,11 +718,11 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
-      'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
-      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
     );
     if (instance!.requestLoading != null) {
       instance.requestLoading!(
@@ -744,14 +744,38 @@ class WebViewClientFlutterApiImpl extends WebViewClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
+      'InstanceManager does not contain a WebViewClient with instanceId: $instanceId',
+    );
+    assert(
+      webViewInstance != null,
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
+    );
+    if (instance!.urlLoading != null) {
+      instance.urlLoading!(webViewInstance!, url);
+    }
+  }
+
+  @override
+  void doUpdateVisitedHistory(
+    int instanceId,
+    int webViewInstanceId,
+    String url,
+    bool isReload,
+  ) {
+    final WebViewClient? instance = instanceManager
+        .getInstanceWithWeakReference(instanceId) as WebViewClient?;
+    final WebView? webViewInstance = instanceManager
+        .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
+    assert(
+      instance != null,
       'InstanceManager does not contain an WebViewClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
       'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
     );
-    if (instance!.urlLoading != null) {
-      instance.urlLoading!(webViewInstance!, url);
+    if (instance!.doUpdateVisitedHistory != null) {
+      instance.doUpdateVisitedHistory!(webViewInstance!, url, isReload);
     }
   }
 
@@ -822,7 +846,7 @@ class DownloadListenerFlutterApiImpl extends DownloadListenerFlutterApi {
         .getInstanceWithWeakReference(instanceId) as DownloadListener?;
     assert(
       instance != null,
-      'InstanceManager does not contain an DownloadListener with instanceId: $instanceId',
+      'InstanceManager does not contain a DownloadListener with instanceId: $instanceId',
     );
     instance!.onDownloadStart(
       url,
@@ -882,11 +906,11 @@ class WebChromeClientFlutterApiImpl extends WebChromeClientFlutterApi {
         .getInstanceWithWeakReference(webViewInstanceId) as WebView?;
     assert(
       instance != null,
-      'InstanceManager does not contain an WebChromeClient with instanceId: $instanceId',
+      'InstanceManager does not contain a WebChromeClient with instanceId: $instanceId',
     );
     assert(
       webViewInstance != null,
-      'InstanceManager does not contain an WebView with instanceId: $webViewInstanceId',
+      'InstanceManager does not contain a WebView with instanceId: $webViewInstanceId',
     );
     if (instance!.onProgressChanged != null) {
       instance.onProgressChanged!(webViewInstance!, progress);