+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/WebChromeClientFlutterApiImpl.java
@@ -32,7 +32,7 @@ public class WebChromeClientFlutterApiImpl extends WebChromeClientFlutterApi {
    * @param instanceManager maintains instances stored to communicate with Dart objects
    */
   public WebChromeClientFlutterApiImpl(
-      BinaryMessenger binaryMessenger, InstanceManager instanceManager) {
+      @NonNull BinaryMessenger binaryMessenger, @NonNull InstanceManager instanceManager) {
     super(binaryMessenger);
     this.binaryMessenger = binaryMessenger;
     this.instanceManager = instanceManager;
@@ -41,7 +41,10 @@ public class WebChromeClientFlutterApiImpl extends WebChromeClientFlutterApi {
 
   /** Passes arguments from {@link WebChromeClient#onProgressChanged} to Dart. */
   public void onProgressChanged(
-      WebChromeClient webChromeClient, WebView webView, Long progress, Reply<Void> callback) {
+      @NonNull WebChromeClient webChromeClient,
+      @NonNull WebView webView,
+      @NonNull Long progress,
+      @NonNull Reply<Void> callback) {
     webViewFlutterApi.create(webView, reply -> {});
 
     final Long webViewIdentifier =
@@ -53,10 +56,10 @@ public class WebChromeClientFlutterApiImpl extends WebChromeClientFlutterApi {
   /** Passes arguments from {@link WebChromeClient#onShowFileChooser} to Dart. */
   @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
   public void onShowFileChooser(
-      WebChromeClient webChromeClient,
-      WebView webView,
-      WebChromeClient.FileChooserParams fileChooserParams,
-      Reply<List<String>> callback) {
+      @NonNull WebChromeClient webChromeClient,
+      @NonNull WebView webView,
+      @NonNull WebChromeClient.FileChooserParams fileChooserParams,
+      @NonNull Reply<List<String>> callback) {
     webViewFlutterApi.create(webView, reply -> {});
 
     new FileChooserParamsFlutterApiImpl(binaryMessenger, instanceManager)