+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/WebChromeClientHostApiImpl.java
@@ -47,7 +47,7 @@ public class WebChromeClientHostApiImpl implements WebChromeClientHostApi {
     }
 
     @Override
-    public void onProgressChanged(WebView view, int progress) {
+    public void onProgressChanged(@NonNull WebView view, int progress) {
       flutterApi.onProgressChanged(this, view, (long) progress, reply -> {});
     }
 
@@ -63,11 +63,12 @@ public class WebChromeClientHostApiImpl implements WebChromeClientHostApi {
     }
 
     @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
+    @SuppressWarnings("LambdaLast")
     @Override
     public boolean onShowFileChooser(
-        WebView webView,
-        ValueCallback<Uri[]> filePathCallback,
-        FileChooserParams fileChooserParams) {
+        @NonNull WebView webView,
+        @NonNull ValueCallback<Uri[]> filePathCallback,
+        @NonNull FileChooserParams fileChooserParams) {
       final boolean currentReturnValueForOnShowFileChooser = returnValueForOnShowFileChooser;
       flutterApi.onShowFileChooser(
           this,
@@ -98,11 +99,14 @@ public class WebChromeClientHostApiImpl implements WebChromeClientHostApi {
    * window.
    */
   public static class SecureWebChromeClient extends WebChromeClient {
-    @Nullable private WebViewClient webViewClient;
+    @Nullable WebViewClient webViewClient;
 
     @Override
     public boolean onCreateWindow(
-        final WebView view, boolean isDialog, boolean isUserGesture, Message resultMsg) {
+        @NonNull final WebView view,
+        boolean isDialog,
+        boolean isUserGesture,
+        @NonNull Message resultMsg) {
       return onCreateWindow(view, resultMsg, new WebView(view.getContext()));
     }
 
@@ -121,7 +125,9 @@ public class WebChromeClientHostApiImpl implements WebChromeClientHostApi {
      */
     @VisibleForTesting
     boolean onCreateWindow(
-        final WebView view, Message resultMsg, @Nullable WebView onCreateWindowWebView) {
+        @NonNull final WebView view,
+        @NonNull Message resultMsg,
+        @Nullable WebView onCreateWindowWebView) {
       // WebChromeClient requires a WebViewClient because of a bug fix that makes
       // calls to WebViewClient.requestLoading/WebViewClient.urlLoading when a new
       // window is opened. This is to make sure a url opened by `Window.open` has
@@ -142,9 +148,8 @@ public class WebChromeClientHostApiImpl implements WebChromeClientHostApi {
               return true;
             }
 
-            // Legacy codepath for < N.
-            @SuppressWarnings("deprecation")
             @Override
+            @SuppressWarnings({"deprecation", "RedundantSuppression"})
             public boolean shouldOverrideUrlLoading(WebView windowWebView, String url) {
               if (!webViewClient.shouldOverrideUrlLoading(view, url)) {
                 view.loadUrl(url);
@@ -184,7 +189,9 @@ public class WebChromeClientHostApiImpl implements WebChromeClientHostApi {
      * @param flutterApi handles sending messages to Dart
      * @return the created {@link WebChromeClientHostApiImpl.WebChromeClientImpl}
      */
-    public WebChromeClientImpl createWebChromeClient(WebChromeClientFlutterApiImpl flutterApi) {
+    @NonNull
+    public WebChromeClientImpl createWebChromeClient(
+        @NonNull WebChromeClientFlutterApiImpl flutterApi) {
       return new WebChromeClientImpl(flutterApi);
     }
   }
@@ -197,16 +204,16 @@ public class WebChromeClientHostApiImpl implements WebChromeClientHostApi {
    * @param flutterApi handles sending messages to Dart
    */
   public WebChromeClientHostApiImpl(
-      InstanceManager instanceManager,
-      WebChromeClientCreator webChromeClientCreator,
-      WebChromeClientFlutterApiImpl flutterApi) {
+      @NonNull InstanceManager instanceManager,
+      @NonNull WebChromeClientCreator webChromeClientCreator,
+      @NonNull WebChromeClientFlutterApiImpl flutterApi) {
     this.instanceManager = instanceManager;
     this.webChromeClientCreator = webChromeClientCreator;
     this.flutterApi = flutterApi;
   }
 
   @Override
-  public void create(Long instanceId) {
+  public void create(@NonNull Long instanceId) {
     final WebChromeClient webChromeClient =
         webChromeClientCreator.createWebChromeClient(flutterApi);
     instanceManager.addDartCreatedInstance(webChromeClient, instanceId);