+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Adds support for `PlatformNavigationDelegate.onUrlChange`.
 * Bumps androidx.webkit:webkit from 1.6.0 to 1.6.1.
+* Fixes common typos in tests and documentation.
 
 ## 3.4.5